levelDropedIten in enemies

+++ Platformer/Assets/Scripts/Items/Data/StatChangingItemDescriptor.cs
@@ -10,10 +10,10 @@ namespace Items.Data
     [Serializable]
     public class StatChangingItemDescriptor : ItemDescriptor
     {
-        [field: SerializeField] public float Level { get; private set; }
+        [field: SerializeField] public int Level { get; private set; }
         [field: SerializeField] public List<StatModificator> Stats { get; private set; }
 
-        public StatChangingItemDescriptor(ItemId itemId, ItemType type, Sprite itemSprite, ItemRarity itemRarity, float price, float level, List<StatModificator> stats):
+        public StatChangingItemDescriptor(ItemId itemId, ItemType type, Sprite itemSprite, ItemRarity itemRarity, float price, int level, List<StatModificator> stats):
             base(itemId, type, itemSprite, itemRarity, price)
         {
             Level = level;
